pyproject.toml
[project]
name = "test-case-wallet"
version = "0.1.0"
description = ""
authors = [
    {name = "Acejkee",email = "belousow.ilja@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (==4.2)",
    "djangorestframework (==3.14)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
